requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.24
chardet==3.0.4
Click==7.0
coverage==4.5.1
Django==2.1.2
django-celery==3.2.2
django-redis==4.9.0
django-test-addons==1.0
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
dnspython==1.15.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
kombu==3.0.37
monotonic==1.5
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyee==5.0.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
unicodecsv==0.14.1
urllib3==1.23
